// src/app/PruebasApi/model/filteredComputer.ts
/**
 * AdminApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1.0.7
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TimeSpan } from './timeSpan';
import { ComputerDetailedStatusDto } from './computerDetailedStatusDto';
import { ComputerStatusDto } from './computerStatusDto';


export interface FilteredComputer { 
    partnerId?: number;
    partnerName?: string | null;
    userId?: number;
    email?: string | null;
    licenseId?: string | null;
    computerId?: number;
    machineName?: string | null;
    alias?: string | null;
    creationDate?: string;
    cancelledDate?: string | null;
    subscriptionExpirationDate?: string | null;
    lastConnection?: string | null;
    protectedDate?: string | null;
    isExpired?: boolean;
    operatingSystem?: string | null;
    motherboard?: string | null;
    processor?: string | null;
    numberOfDrives?: number;
    partitions?: string | null;
    drivesTotalLength?: number;
    status?: ComputerStatusDto;
    detailedStatus?: ComputerDetailedStatusDto;
    timeSinceLastConnection?: TimeSpan;
    readonly timeSinceExpiration?: number | null;
}
export namespace FilteredComputer {
}