// src/app/PruebasApi/model/getPartnerComputersRequest.ts
/**
 * AdminApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1.0.7
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GetComputersFilter } from './getComputersFilter';
import { GetItemsRequestOrder } from './getItemsRequestOrder';
import { GetItemsRequestPaging } from './getItemsRequestPaging';


export interface GetPartnerComputersRequest {
    filter?: GetComputersFilter;
    order?: GetItemsRequestOrder;
    paging?: GetItemsRequestPaging;
}